Rec: Move RecognizeRunner to dedicated namespace
