+++ PatternPal/PatternPal/Services/StepByStepService.cs
@@ -1,5 +1,6 @@
 ﻿#region
 
+using PatternPal.Core.Runner;
 using PatternPal.Core.StepByStep;
 
 using InstructionSet = PatternPal.Protos.InstructionSet;
